Update kubernetes guide
Made a few small edits - first the name "external client" is a bit confusing and not quite obvious what the guide is about (it's not about client, it's about connecting to hazelcast). The guide title should have an objective and be action oriented rather than what it's about.
Also whether the client is smart or not, it's not so relevant for this guide - it's too much detail.