Show the note about the `minikube tunnel` before creating the services
In the smart client guide, we have a note regarding the `minikube tunnel`,
but it is only shown after the services are created.

But, executing this command after the services are created will not
be effective, as the members will not be able to resolve their public
IPs (as they are not assigned yet, since the `tunnel` is not started)
and the client will not be able to connect cluster members, apart
from the initial connection.

This PR fixes that problem by moving the note to the top, before we
create the services for the smart client case.